requirements/base.txt
Django==4.0.4
django-filter==21.1
djangorestframework==3.13.1
pandas==1.4.2
psycopg2==2.9.3
matplotlib==3.5.1
pandas-datareader==0.10.0
tensorflow==2.8.0
keras==2.8.0
scikit-learn==1.0.2
beautifulsoup4==4.11.1
selenium==4.1.3
mplfinance==0.12.8b9
